move construct_supernode_feature() to common.py
